requirements.txt
# AI History Video Generator - Dependencies
# UPSC & State PSC History Course

# LLM Integration
groq>=0.4.0
ollama>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.10

# Text-to-Speech
edge-tts>=6.1.0
pydub>=0.25.1

# PDF Notes Generation
reportlab>=4.0.0

# Avatar Generation
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0
librosa>=0.10.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Video Processing
moviepy==1.0.3
decorator==4.4.2
Pillow>=10.0.0
ffmpeg-python>=0.2.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Utilities
pyyaml>=6.0
apscheduler>=3.10.0
SQLAlchemy>=2.0.0
tqdm>=4.66.0
loguru>=0.7.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0
pytz>=2023.3

# Face Enhancement (optional - for better avatar quality)
# gfpgan>=1.3.0  # Uncomment if GPU available

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
